added pojos

+++ src/main/java/com/demo/PlayerMicroService/Repository/PlayerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
-public interface PlayerRepository extends JpaRepository<Player, Long> {
-    Optional<Player> findById(Long id);
+public interface PlayerRepository extends JpaRepository<Player, Integer> {
+    Optional<Player> findById(int id);
 }
